refactor: break the multi_agent file into the initial structure
